requirements.txt
flask-unchained>=0.1
flask-controller-bundle>=0.1
flask-marshmallow>=0.8.0
flask-sqlalchemy-bundle>=0.1
marshmallow>=2.13.6
marshmallow-sqlalchemy>=0.13.1
